+++ R/plot_rawEMG.R
@@ -9,16 +9,13 @@
 #' @param dark_mode To enable dark mode
 #' @param line_size Line thickness
 #' @param line_col Line colour
-#' @param path_for_graphs Path where plots should be saved
-#' @param filetype Plot file type (e.g., "png" or "svg")
-#' @param width Plot width in pixels
-#' @param height Plot height in pixels, for each muscle
-#' @param resolution Plot resolution in pixels
+#' @param show_plot Logical, to decide whether plots should be plotted in the active graphic device
 #'
 #' @details
-#' If `path_for_graphs` is not specified, plots will appear in the plot pane and will not be saved.
+#' If `show_plot` is TRUE (default) plots are also shown in the active graphic device.
+#' Plots can then be saved with the preferred export method, such as `ggplot2::ggsave`.
 #'
-#' @return Exports raw EMG plots of the specified length.
+#' @return Plots raw EMG trials of the specified length.
 #'
 #' @export
 #'
@@ -26,17 +23,14 @@
 #' # Load some data
 #' data(RAW_DATA)
 #'
-#' # Plot
-#' for (ii in seq_along(RAW_DATA)) {
-#'   # Plot the first three seconds of each trial in dark mode
-#'   plot_rawEMG(RAW_DATA[[ii]],
-#'     trial = names(RAW_DATA)[ii],
-#'     row_number = 4,
-#'     col_number = 4,
-#'     dark_mode = TRUE,
-#'     line_col = "tomato3"
-#'   )
-#' }
+#' # Plot first (and only) trial in RAW_DATA, first three seconds, in dark mode
+#' plot_rawEMG(RAW_DATA[[1]],
+#'   trial = names(RAW_DATA)[1],
+#'   row_number = 4,
+#'   col_number = 4,
+#'   dark_mode = TRUE,
+#'   line_col = "tomato3"
+#' )
 plot_rawEMG <- function(x,
                         trial,
                         plot_time = 3,
@@ -46,11 +40,7 @@ plot_rawEMG <- function(x,
                         dark_mode = FALSE,
                         line_size = 0.3,
                         line_col = "black",
-                        path_for_graphs = NA,
-                        filetype,
-                        width,
-                        height,
-                        resolution) {
+                        show_plot = TRUE) {
   if (!inherits(x, "EMG")) {
     stop("Object is not of class EMG, please create objects in the right format with \"rawdata\"")
   } else {
@@ -81,19 +71,6 @@ plot_rawEMG <- function(x,
   } else {
     bg_col <- "white"
     text_col <- "black"
-  }
-
-  # Prepare plot export if needed
-  if (!is.na(path_for_graphs)) {
-    # For exporting
-    Cairo::Cairo(
-      file = paste0(path_for_graphs, trial, ".", filetype),
-      width = width, height = height, dpi = resolution, bg = bg_col
-    )
-  } else {
-    # For plotting directly (e.g. within RStudio)
-    graphics::par(bg = bg_col)
-    graphics::plot.new()
   }
 
   # Create plots
@@ -141,14 +118,20 @@ plot_rawEMG <- function(x,
   # Check for consistency of row and column number
   if (is.na(row_number)) row_number <- ceiling(length(colnames(x)) / col_number)
 
-  # Arrange plots nicely
-  gridExtra::grid.arrange(
+  # Arrange plots nicely and return as gtable
+  gg <- gridExtra::arrangeGrob(
     grobs = varlist,
     nrow = row_number,
     ncol = col_number,
-    top = grid::textGrob(paste0("Trial: ", trial), gp = grid::gpar(col = text_col)),
-    newpage = FALSE
+    top = grid::textGrob(paste0("Trial: ", trial), gp = grid::gpar(col = text_col))
   )
-
-  if (!is.na(path_for_graphs)) grDevices::dev.off()
+  # Plot on active graphic device if needed
+  if (show_plot) {
+    # Prepare graphic device
+    graphics::par(bg = bg_col)
+    graphics::plot.new()
+    # Arrange
+    gridExtra::grid.arrange(gg, newpage = FALSE)
+  }
+  return(gg)
 }